// src/app/shared/models/action-type.ts
export enum ActionType {
  Abandon = 'abandon',
  Author = 'author',
  Cancel = 'cancel',
  ChangeHistory = 'changeHistory',
  Close = 'close',
  Comments = 'comments',
  Confirmation = 'confirm',
  Editor = 'editor',
  No = 'no',
  Ok = 'ok',
  Pause = 'pause',
  Review = 'review',
  Save = 'save',
  Skip = 'skip',
  Yes = 'yes',
  GotIt = 'Got it',
  Unlock = 'Unlock'
}