// src/app/shared/models/activity-status.ts
export enum ActivityStatus {
  Abandoned = 'abandoned',
  // Archived = 'archived',
  Deleted = 'deleted',                // legacy purposes
  // Expired = 'expired',                // legacy purposes
  Issued = 'issued',
  NotSet = 'notSet',
  Published = 'published',
  // Released = 'released',
  Returned = 'returned',
  ReadyToReview = 'readyToReview',
  ReviewApproved = 'reviewApproved',
  UnderAmendment = 'underAmendment',
  UnderConstruction = 'draft',
  UnderReview = 'underReview',
  Withdrawn = 'withdrawn',
}